Add support for running against a container
When running against a container, specifically Testcontainers, the advertised address will be on the standard port, 6648. However the port exposed to the client on the host will be a random mapped port. In this case we can intercept the shard manager and just use the original standalone service address instead.

This also fixes a couple of minor warnings - compiler unchecked warnings and missing slf4j implementation for tests.